requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
black==20.8b1
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
datadog==0.39.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.4
django-json-widget==1.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
entrypoints==0.3
flake8==3.8.3
Flask==1.1.2
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
packaging==20.7
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==20.0.0
qrcode==6.1
regex==2020.7.14
requests==2.25.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
